# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Textile ERP API - نظام إدارة رولات القماش
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new tenant and admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /fabric-rolls:
    get:
      operationId: listFabricRolls
      tags: [fabric-rolls]
      summary: List fabric rolls
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: color
          in: query
          schema:
            type: string
        - name: gsm
          in: query
          schema:
            type: number
        - name: productionOrderId
          in: query
          schema:
            type: integer
        - name: warehouseId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of fabric rolls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FabricRoll"
  /fabric-rolls/{id}:
    get:
      operationId: getFabricRoll
      tags: [fabric-rolls]
      summary: Get fabric roll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fabric roll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FabricRoll"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFabricRoll
      tags: [fabric-rolls]
      summary: Update fabric roll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFabricRollRequest"
      responses:
        "200":
          description: Updated fabric roll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FabricRoll"
  /fabric-rolls/by-code/{rollCode}:
    get:
      operationId: getFabricRollByCode
      tags: [fabric-rolls]
      summary: Get fabric roll by roll code (for QR scan)
      parameters:
        - name: rollCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Fabric roll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FabricRoll"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # QC Reports
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        companyName:
          type: string
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
      required:
        - companyName
        - email
        - password
        - fullName
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tenantId
        - email
        - fullName
        - role
        - isActive
        - createdAt
        - updatedAt
...
    FabricRoll:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        rollCode:
          type: string
        batchId:
          type: string
        productionOrderId:
          type: integer
        warehouseId:
          type: ["integer", "null"]
        length:
          type: number
        weight:
          type: number
        color:
          type: string
        gsm:
          type: number
        width:
          type: number
        fabricType:
          type: string
        status:
          type: string
        qrCode:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tenantId
        - rollCode
        - batchId
        - productionOrderId
        - length
        - weight
        - color
        - gsm
        - width
        - fabricType
        - status
        - qrCode
        - createdAt
        - updatedAt
    UpdateFabricRollRequest:
      type: object
      properties:
        status:
          type: string
        warehouseId:
          type: integer
        color:
          type: string
        length:
          type: number
        weight:
          type: number
        notes:
          type: string
...